+++ facade/server/config.js
@@ -60,25 +60,9 @@ module.exports = {
         chatId: process.env.TELEGRAM_CHAT_ID
     },
     
-    fileBrowser: {
-        uploadPath: path.resolve(process.env.FILEBROWSER_UPLOAD_PATH),
-        rootPath: path.resolve(process.env.FILEBROWSER_ROOT_PATH)
-    },
-    
     gallery: {
-        // currently only `local` gallery storage is supported
-        storage: process.env.GALLERY_STORAGE,
-        uploadPath: path.resolve(process.env.GALLERY_UPLOAD_PATH),
-        rootPath: path.resolve(process.env.GALLERY_ROOT_PATH),
-        photosPath: process.env.GALLERY_PHOTOS_PATH,
-        photoSetsPath: process.env.GALLERY_PHOTOSETS_PATH,
-        trashPath: process.env.GALLERY_TRASH_PATH,
-        // image processing module (sharp or jimp, must be installed)
-        imageProcessingModule: process.env.GALLERY_IMAGE_PROCESSING_MODULE,
-        jpgQuality: parseInt(process.env.GALLERY_JPG_QUALITY),
-        // comma-separated image sizes: "<suffix>:<width>x<height>"
-        imageSizes: process.env.GALLERY_IMAGE_SIZES,
-        dashboardThumbnailSuffix: process.env.GALLERY_DASHBOARD_THUMBNAIL_SUFFIX,
-        newPhotosFirst: process.env.GALLERY_NEW_PHOTOS_FIRST === "true"
+        newPhotosFirst: process.env.GALLERY_NEW_PHOTOS_FIRST === "true",
+        defaultPhotoThumbnailSuffix: process.env.GALLERY_DEFAULT_PHOTO_THUMBNAIL_SUFFIX,
+        defaultPhotoSuffix: process.env.GALLERY_DEFAULT_PHOTO_SUFFIX,
     }
 };