+++ facade/server/lib/utils.js
@@ -10,6 +10,14 @@ function partial(fn) {
     };
 }
 
+function addFilenameSuffix(filename, suffix) {
+    const path = require("path"),
+        extension = path.extname(filename);
+    filename = filename.slice(0, -extension.length);
+    return filename + "_" + suffix + extension;
+}
+
 module.exports = {
-    partial: partial
+    partial,
+    addFilenameSuffix
 };