+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * resize photoset cover;
     * `jimp` support dropped;
     * refresh photoset's cover after changing it;
+    * view photo set - load the smallest thumbnails by default;
 * API - `REGISTRATION_MODE` option added (possible values are `open` and `invite`);
 
 ## **2.0.1** - *2020-04-01*